Extract desaparicion detection
