+++ backend/src/modules/sessionMemory.js
@@ -83,20 +83,4 @@ export async function actualizarUltimaHora(userId, timestamp = new Date()) {
   }
 }
 
-export async function detectarDesaparicion(userId, minutos = 5) {
-  try {
-    const session = await Session.findOne({ userId }).exec();
-    if (!session || !session.ultimaRespuestaHora) return false;
-    const diffMin =
-      (Date.now() - new Date(session.ultimaRespuestaHora).getTime()) / 60000;
-    const desaparecido = diffMin > minutos;
-    if (session.desaparecido !== desaparecido) {
-      session.desaparecido = desaparecido;
-      await session.save();
-    }
-    return desaparecido;
-  } catch (err) {
-    console.error('Error detectando desaparicion:', err);
-    return false;
-  }
-}
+export { detectarDesaparicion } from './detectorDesaparicion.js';
